# docs/swagger.yaml
definitions:
...
  dto.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginByPin:
    properties:
      pin:
        type: string
    required:
    - pin
    type: object
...
  dto.RefreshToken:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  middlewares.ResponseError:
    properties:
      message: {}
      request_id:
        type: string
      time:
        type: string
    type: object
  models.Action:
    properties:
      api:
        type: string
      created_at:
        type: string
      default:
        type: boolean
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ActionsMenus:
    properties:
      actionId:
        type: integer
      actions:
        $ref: '#/definitions/models.Action'
      created_at:
        type: string
      id:
        type: integer
      menuId:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Menu:
    properties:
      actions_menus:
        items:
          $ref: '#/definitions/models.ActionsMenus'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      object_menus:
        items: {}
        type: array
      roles_menus:
        items:
          $ref: '#/definitions/models.RolesMenus'
        type: array
      updated_at:
        type: string
    type: object
  models.RolesMenus:
    properties:
      created_at:
        type: string
      id:
        type: integer
      menu:
        $ref: '#/definitions/models.Menu'
      menu_id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      enable_login_by_link:
        type: boolean
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      phone:
        type: string
      pin:
        type: string
      profile_pict:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  services.LoginResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
...
info:
  contact:
    name: PT Tata Kelola Dokumen
  description: Aplikasi Assessment & Monitoring
  license:
    name: Private License
  title: EJM Assessment & Monitoring API Documentation
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bahasa
        enum:
        - en
        - id
        in: header
        name: Accept-Language
        type: string
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Autentikasi User untuk Resource Private Server
      tags:
      - Auth
  /auth/oauth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bahasa
        enum:
        - en
        - id
        in: header
        name: Accept-Language
        type: string
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginByPin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Autentikasi User untuk Resource Private Server
      tags:
      - Auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bahasa
        enum:
        - en
        - id
        in: header
        name: Accept-Language
        type: string
      - description: Refresh Token
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshToken'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Refresh JWT Token Yang Expired
      tags:
      - Auth
...
  /session:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bahasa
        enum:
        - en
        - id
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Data Sesi
      tags:
      - Auth
...
schemes:
- http
swagger: "2.0"